Fix unrecognized path have 404
